moved HDF5 to root folder
