+++ nac/schedule/__init__.py
@@ -2,9 +2,8 @@ from .scheduleCoupling import (compute_phases, lazy_couplings,
                                write_hamiltonians)
 from .scheduleCp2k import (prepare_cp2k_settings, prepare_job_cp2k)
 from .components import (calculate_mos, create_point_folder, split_file_geometries)
-from .hdf5_interface import StoreasHDF5
 
-__all__ = ['StoreasHDF5', 'calculate_mos', 'compute_phases',
+__all__ = ['calculate_mos', 'compute_phases',
            'create_point_folder', 'lazy_couplings',
            'prepare_cp2k_settings', 'prepare_job_cp2k',
            'store_cp2k_basis', 'split_file_geometries', 'write_hamiltonians']
